회원가입과 유저 관리 객체를 분리
